requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pillow
requests
httpx
psycopg2-binary
streamlit
exifread
folium
streamlit-folium
geopy
numpy
python-dotenv
tensorflow
huggingface-hub